Fix generated links for DiffCommit resources
Previously, the generated links for the `DiffCommitResource` and
`DraftDiffCommitResource` were being overridden by the resources for
the parent `DiffResource` or `DraftDiffResource`, respectively. Now,
only the `files` link is taken from the parent resource and the other
links are not overriden

Testing Done:
Ran unit test.

Checked the links of a `DiffCommitResource` and saw that they were
correct.

Reviewed at https://reviews.reviewboard.org/r/7132/