Allow fetching patch files for individual commits
The patch file for an individual commit is now fetchable by sending the
`Accept: text/x-patch` header to the `DiffCommitResource` and
`DraftDiffCommitResource`. This patch is composed of the individual
patches for all `FileDiff`s in the `DiffCommit`.

Testing Done:
Ran unit tests.

Uploaded a review request with commit history and made an HTTP request
to the `DiffCommitResource` API endpoint with a `Accept: text/x-patch`
header. The returned response was a patch file that contained only the
diff information for `FileDiff`s in that commit.

Reviewed at https://reviews.reviewboard.org/r/7131/