Switch spinners to use Font Awesome's fa-spinner.
Our old spinner.gif is one of the ugliest remaining parts of Review Board. It's
not retina-aware, and it feels pretty out of place in the modern world.

This change swaps out most of the uses of spinner.gif to use Font Awesome's
fa-spinner icon. This uses CSS transforms to rotate the image, and scales
cleanly to any size and resolution.

There are a couple more uses of the old spinner (in rich text rendering and the
new review request/upload diff UIs) which I have yet to clean up. I'll get to
those in a later change.

Testing done:
Looked at each piece of UI (often by making manual changes to the code to
ensure that the spinner stayed on screen long enough to really verify it).

Reviewed at https://reviews.reviewboard.org/r/7173/