Reduce spurious logging for file attachments.
In cases where a mimetype handler or review UI rejects a given file attachment,
we log the output. This is useful for knowing, for example, why the PDF review
UI isn't trying to render for something which has mimetype application/pdf but
is actually a .docx file. However, in this case we were also logging with
`exc_info` turned on, which filled the log file with useless backtraces. This
change removes those.

Reviewed at https://reviews.reviewboard.org/r/7174/