Upload Attachment View - Submit form without specifying a file
Currently, if you open the modal for file uploads, you would be presented a form for uploads. Currently, the upload button is enabled, and submission is allowed without specifying a file to submit. This patch would allow set the button to disabled by default, and only enable it if a file is specified.

Testing Done:
Manual testing was performed.

Bugs closed: 3829

Reviewed at https://reviews.reviewboard.org/r/7149/