Attach parent diffs with renamed and modified files
Previously, if a review request was uploaded with a parent diff it was
possible that parent diffs would not be attached to `FileDiff`s in the
case where the parent diff for a file was the result of a rename and a
modify. This would result in the underlying SCM not being able to
locate the parent revision and generate the original file for the
diffviewer.

Now instead of attaching parent diffs to `FileDiff`s based on the
original file names in the parent diff, we consider the destination
file names. This results in parent diffs being correctly attached when
they contain a renamed and modified file.

Testing Done:
Ran unit tests.

Uploaded a review request with a parent diff that had a modified and
renamed file. The diff was visible in the diffviewer when uploaded
with this patch applied; it was not visible when it was uploaded
without the patch applied.

Reviewed at https://reviews.reviewboard.org/r/7178/