Require Django Haystack 2.3.1 at a minimum.
We had a report of Haystack 2.3.0 failing to index, due to the top-level
djblets module not having a models.py. 2.3.1 fixes this. We now require
at least this version.

Testing Done:
The original reporter verified that this fixed the problem.

Built the package.

Bugs closed: 3836

Reviewed at https://reviews.reviewboard.org/r/7179/