Fix issues when searching for strings of digits.
We have an escape hatch that allows users to type in a review request ID into
the search box and jump straight there. Unfortunately, the way that it was
implemented meant that if there were no results for that (say, a user was
typing in a bug number instead), it would crash.

It turns out that we already had code that handled looking up the review
request ID directly. I've added some additional permissions checking to that to
make us only redirect when the searching user has access to the review request,
and otherwise the search is done using the text rather than trying to have
special cases for digit searches vs. non-digit searches.

Testing done:
- Ran unit tests.
- Searched for a string of digits that did not correspond to a review request
  ID.

Fixes bug 3820.

Reviewed at https://reviews.reviewboard.org/r/7185/