Fix line counts returned by DiffCommitModel
Previously, the total line counts returned was the concatenation of the
keys of the `lineCounts` object. This has been fixed to return the
actual sum of the elements in the object and the logic has been
simplified. This fixes the visual style of the
`RB.DiffComplexityIconView`s in the `RB.DiffCommitIndexView`.

Testing Done:
Verified the line counts returned by the `getTotalLineCounts` function
were indeed an integer.

Reviewed at https://reviews.reviewboard.org/r/7236/