Ensure interdiffs do not update line counts
Previously it was possible for the `DiffChunkGenerator` to update the
line counts in a forced interdiff (i.e., an interdiff where there is no
corresponding interfilediff, such as when the changes to a file are
reverted). This results in all insert and delete counts being swapped
in the `FileDiff`'s line counts. This patch fixes this issue.

A unit test was added to ensure that this does not regress.

Testing Done:
Ran unit tests.
The new unit test failed without the patch applied.

Reviewed at https://reviews.reviewboard.org/r/7241/