Give all datagrid pages access to some dashboard-level functionality.
The dashboard page was the only one that could auto-size to fit the
page, auto-reload content from the server, or set up any multi-select
actions. This was all due to the `DashboardView` class powering it in
JavaScript.

That class has since been pulled out into a more reusable
`DatagridPageView` and accompanying `DatagridPageModel`. These control the
auto-sizing, and allow datagrids to take advantage of actions and
periodic reloads, along with other events.

The dashboard is still the only one with multi-select actions, but all
the other pages now auto-fit, which looks far nicer than what we had
before. In time, we may decide to port other features to these
datagrids, now that we can do so more easily.

Since all datagrids now fit the dimensions of the page, the special
class for a scrollable sidebar has been removed. That mode is now the
default.

Testing Done:
Tested each datagrid page we have. Saw that they were all nicely fitting
to the page dimensions.

Tested that the sidebars were doing the right thing on the dashboard and
user pages.

Tested dashboard actions.

Reviewed at https://reviews.reviewboard.org/r/7225/