Improve the styling of the summary state labels in the datagrids.
The summary labels ("Draft", "Submitted", "Discarded") were just simple
text prefixes before, which were a bit ugly with the new datagrids, and
didn't do a great job of standing out.

This adds some styling to those to make them appear as little labels.
They're placed to the left of the text block, and each have their own
distinct color.

The summary text retains its own solid block. It no longer wraps below
the label, but instead stays to the right of it. It also has some
additional line spacing, helping to improve readability when it does
wrap.

I also changed the "No Summary" text to be slightly faded out, so that
it doesn't blend in with actual summaries as easily.

Testing Done:
Tested the labels at various widths, to ensure the wrapping was correct
an the default state was correct.

Reviewed at https://reviews.reviewboard.org/r/7228/