Clean up the header of the page, and theme it for specific pages.
In 2.0, the header at the top of the page was separated into the top
black section with "Review Board," the version, search, and links; and the
bottom blue section with navigation.

This changes the header to be a single clean entity designed to be a bit
less noisy, visually, and to fit in better with the given page. The
colors on the bar are no longer exactly the same across all pages in
Review Board, but rather are designed to fit in with the color palette
already on those pages. For instance, on the dashboard or My Account
pages, it uses a light blue color scheme, while review requests use a
brown scheme.

Not all pages have been updated for this yet. The administration UI has
a sort of boring grey for now, for instance. I'll be updating these as I
continue with the UI redesign.

Note also that the feel of this on the review request page will be
improving along with the new redesign of that page, coming soon to a
theater near you.

Testing Done:
Checked all the pages I could think of. Saw a suitable color scheme for the bar.

I also checked the menus for each page, which had suitable colors as well.

Reviewed at https://reviews.reviewboard.org/r/7238/