Fix a UnicodeWarning comparing bytes with unicode in the condensediff command.
In the condensediff management command on release-2.0.x, at one point we were
comparing bytes with unicode, which will always fail. This resulted in a
spurious warning and extra work. Fixing the empty string to be an empty
bytestring fixes the warning.

Based on Rohan Meringenti's work at /r/6949.

Fixes bug 3698.