Add bug_tracker to available GET fields for a repository.
This can be useful for e.g. the hook review in request
https://reviews.reviewboard.org/r/7100/, which (for now) has to configure the
bug tracker URL in the Mercurial config (hgrc) in addition to the Review Board
configuration.

Testing Done:
Tested that the `bug_tracker` can be read by the RBTools Python interface to
the web API.

Reviewed at https://reviews.reviewboard.org/r/7255/