Rework archive/mute UI.
The UI for archive and mute within a review request was a little bit intrusive,
and given how complicated this page already is, was too cluttered. In
particular, the addition of new banners at the top of the page (which
inadvertantly hid existing banners like the draft banner or closed banner) were
just far too much UI for this.

I've replaced the "Hide" menu and the banners with a new menu on the left-hand
side of the review request box, just next to the star icon. This uses a new
"inbox" icon in the same style as the star icon, which highlights when the
review request is archived or muted.

As part of this, archiving/muting/unarchiving/unmuting no longer reloads the
page. Instead, the code in the ReviewRequestEditorView will update the
visibility of the relevant menu items after the state is saved to the server.
Once we have a better icon, I intend to also have it change the state of the
icon.

Testing done:
- Archived, unarchived, muted, and unmuted review requests.
- Checked that the state was always tracked correctly.
- Ran unit tests.
- Ran js-tests.