Make some of the datagrid classes more generically useful.
The recent dashboard class revamp renamed much of the dashboard-specific
classes to datagrid-specific classes, so we could use them on more
pages. This change goes a step further, taking many of those classes and
making them useful outside of datagrids.

There are now a couple of top-level classes for `<body>` that define
whether the page has a sidebar or whether the content will be taking up
the full visible height of the page.

Some of the datagrid page IDs (`#datagrid_page_container` and
`#datagrid_page_content`) have been removed, instead making of use of
`#page-container`, `#content_container`, and `#content`. That's two layers of
elements and styles that are removed and made generic.

The sidebar classes are also now generic, no longer tied to datagrids.
This will allow us to reuse it for the New Review Request page, and
others.

Testing Done:
Tested all the datagrid pages in desktop and mobile modes. Made sure
fit-to-viewport, scrolling, and resizing all worked as expected.

Tested the review request, diff viewer, and admin pages.

Reviewed at https://reviews.reviewboard.org/r/7260/