Include the file and interdiff information in diff fragment ETags.
The diff fragments were only considering diff revision numbers (by way
of the URL) for the ETag, and not the IDs of the objects in the
database. This resulted in stale caches if trying to upload another
draft of a diff with the same revision.

We now include both the filediff ID and the interdiffset ID (if any).

Testing Done:
Manually tested, and saw that subsequent uploads of a diff were rendering without
having to clear cache.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/7273/