Fix the interdiff URL changing after load in the diff viewer.
When viewing an interdiff, the revision range in the URL would change,
stripping away the interdiff information. This was due to the
`Backbone.history.start` call's URL failing to account for the interdiff.
We now build this correctly to factor this information in.

Testing Done:
Navigated to various interdiff ranges, and never saw the URL strip the
information away.

Reviewed at https://reviews.reviewboard.org/r/7274/