Fix styling of custom pages and extensions that augment the header bar.
Some extensions set up custom pages, which may not have a "default-page"
class on <body>, breaking default styling.

Other extensions, like rbmotd, add UI to the header bar at the top of
the page.  Due to some of the new style changes, this looked pretty bad,
with a double border at the bottom and no border at the top.  It also
didn't appear on mobile.

This fixes both cases. All pages now have the default styling,
regardless of whether the have that class. I've also moved around the
borders and the `display: none` rule for the navbar to fix display of
rbmotd on desktop and mobile UIs.

Testing Done:
Viewed the Power Pack Reports page, and saw the proper page styling.

Saw that rbmotd's banner looked as it did before, and that it apppeared
correctly on mobile.

Reviewed at https://reviews.reviewboard.org/r/7277/