Ensure we always show a retina version of file attachment thumbnails.
Loading of file attachment thumbnails can sometimes be racy with regards to
retina.js. In particular, if we finish loading everything before the window's
onload event fires, retina.js will happily process all the images and show the
at2x versions. If not (and for file attachments created via drag-and-drop), it
never processes them.

This change updates the FileAttachmentThumbnailView to manually create a
RetinaImage, and remove the data-at2x attribute (so that the onload handler
won't process them).

Testing done:
Uploaded new file attachments and saw the high-res version of the thumbnail.

Reviewed at https://reviews.reviewboard.org/r/7284/