Fix gravatars in review replies and the dashboard.
The review replies weren't able to make use of gravatars properly, due
to the addition of the gravatars setting check. The variable being used
there wasn't available to the template, so I've plumbed a suitable
variable through.

The datagrid was showing gravatars all the time, not considering that
setting. It now checks before deciding to render the gravatar.

Testing Done:
Tested that gravatars appeared in the review replies, and that they were
turned off in both locations when turning off the setting.

Reviewed at https://reviews.reviewboard.org/r/7279/